Source code restructuring
